initial setup of the phishing management; still need to update the campains page
